+++ src/Pug/Nodes/MixinBlock.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Pug\Nodes;
-
-class MixinBlock extends Node
-{
-}