+++ src/Pug/Parser/Exception.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Pug\Parser;
-
-class Exception extends \Exception
-{
-}