+++ tests/features/cache.php
@@ -11,10 +11,11 @@ class PugTest extends Pug
         return $this->compilationsCount;
     }
 
-    public function compile($input, $filename = null)
+    public function compileFile($input)
     {
         $this->compilationsCount++;
-        return parent::compile($input, $filename);
+
+        return parent::compileFile($input);
     }
 }
 
@@ -53,7 +54,7 @@ class PugCacheTest extends PHPUnit_Framework_TestCase
     /**
      * Cache from string input
      */
-    public function testStringInputCache()
+    public function testFileCache()
     {
         $dir = sys_get_temp_dir() . '/pug';
         if (file_exists($dir)) {
@@ -66,16 +67,19 @@ class PugCacheTest extends PHPUnit_Framework_TestCase
         } else {
             mkdir($dir);
         }
+        $test = "$dir/test.pug";
+        file_put_contents($test, "header\n  h1#foo Hello World!\nfooter");
         $pug = new PugTest(array(
+            'debug' => false,
             'cache' => $dir
         ));
-        $this->assertSame(0, $pug->getCompilationsCount(), 'Should have done always 2 compilations because the code changed');
         $this->assertSame(0, $pug->getCompilationsCount(), 'Should have done no compilations yet');
-        $pug->render("header\n  h1#foo Hello World!\nfooter");
+        $pug->renderFile($test);
         $this->assertSame(1, $pug->getCompilationsCount(), 'Should have done 1 compilation');
-        $pug->render("header\n  h1#foo Hello World!\nfooter");
+        $pug->renderFile($test);
         $this->assertSame(1, $pug->getCompilationsCount(), 'Should have done always 1 compilation because the code is cached');
-        $pug->render("header\n  h1#foo Hello World?\nfooter");
+        file_put_contents($test, "header\n  h1#foo Hello World2\nfooter");
+        $pug->renderFile($test);
         $this->assertSame(2, $pug->getCompilationsCount(), 'Should have done always 2 compilations because the code changed');
         $this->emptyDirectory($dir);
     }