+++ tests/features/exceptions.php
@@ -6,30 +6,6 @@ use Pug\Pug;
 class EmulateBugException extends \Exception {}
 class OnlyOnceException extends \Exception {}
 
-class ExtendParser extends Parser
-{
-    public function parse()
-    {
-        static $i = 0;
-        if ($i++) {
-            throw new OnlyOnceException("E: Works only once", 1);
-        }
-        parent::parse();
-    }
-}
-
-class IncludeParser extends Parser
-{
-    public function parse()
-    {
-        static $i = 0;
-        if ($i++) {
-            throw new OnlyOnceException("I: Works only once", 1);
-        }
-        parent::parse();
-    }
-}
-
 class PugExceptionsTest extends PHPUnit_Framework_TestCase
 {
     static public function emulateBug()
@@ -38,8 +14,7 @@ class PugExceptionsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Pug\Parser\Exception
-     * @expectedExceptionCode 10
+     * @expectedException \Exception
      */
     public function testDoNotUnderstand()
     {
@@ -47,8 +22,7 @@ class PugExceptionsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Pug\Parser\Exception
-     * @expectedExceptionCode 10
+     * @expectedException \Exception
      */
     public function testDoubleDoubleArrow()
     {
@@ -56,8 +30,7 @@ class PugExceptionsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException \ErrorException
-     * @expectedExceptionCode 29
+     * @expectedException \Exception
      */
     public function testAbsoluteIncludeWithNoBaseDir()
     {
@@ -66,8 +39,7 @@ class PugExceptionsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException Pug\Parser\Exception
-     * @expectedExceptionCode 10
+     * @expectedException \Exception
      */
     public function testCannotBeReadFromPhp()
     {
@@ -75,8 +47,7 @@ class PugExceptionsTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException \ErrorException
-     * @expectedExceptionCode 34
+     * @expectedException \Exception
      */
     public function testUnexpectedValue()
     {