+++ tests/features/expressionLanguage.php
@@ -1,15 +1,6 @@
 <?php
 
-use Pug\Compiler;
 use Pug\Pug;
-
-class ExpressionCompilerTester extends Compiler
-{
-    public function callPhpizeExpression()
-    {
-        return call_user_func_array([$this, 'phpizeExpression'], func_get_args());
-    }
-}
 
 class PugExpressionLanguageTest extends PHPUnit_Framework_TestCase
 {
@@ -17,7 +8,6 @@ class PugExpressionLanguageTest extends PHPUnit_Framework_TestCase
     {
         $pug = new Pug([
             'debug' => true,
-            'singleQuote' => false,
             'expressionLanguage' => 'js',
         ]);
 
@@ -30,36 +20,8 @@ class PugExpressionLanguageTest extends PHPUnit_Framework_TestCase
         $actual = trim($pug->render("- a = '2'\n- b = 4\np=a + b"));
         $this->assertSame('<p>24</p>', $actual);
 
-        $compiler = new ExpressionCompilerTester([
-            'expressionLanguage' => 'js',
-        ]);
-        $actual = trim($compiler->callPhpizeExpression('addDollarIfNeeded', 'array(1)'));
-        $this->assertSame('array(1)', $actual);
-        $actual = trim($compiler->callPhpizeExpression('addDollarIfNeeded', 'a'));
-        $this->assertSame('$a', $actual);
-
         $actual = trim($pug->render("mixin test\n  div&attributes(attributes)\nbody\n  +test()(class='test')"));
-        $this->assertSame('<body><div class="test "></div></body>', $actual);
-    }
-
-    public function testPhpExpression()
-    {
-        $compiler = new ExpressionCompilerTester([
-            'expressionLanguage' => 'php',
-        ]);
-
-        $actual = trim($compiler->callPhpizeExpression('addDollarIfNeeded', 'a'));
-        $this->assertSame('a', $actual);
-    }
-
-    public function testAutoExpression()
-    {
-        $compiler = new ExpressionCompilerTester([
-            'expressionLanguage' => 3.123,
-        ]);
-
-        $actual = trim($compiler->callPhpizeExpression('addDollarIfNeeded', 'a'));
-        $this->assertSame('$a', $actual);
+        $this->assertSame('<body><div class="test"></div></body>', $actual);
     }
 
     public function testJsLanguageOptions()